backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.1
python-dotenv==1.0.0
google-generativeai==0.3.2
openai==1.6.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
